+++ src/clj_kondo/impl/namespace.clj
@@ -513,6 +513,22 @@
                         (first existing)
                         (str/join ", " (sort existing))))))))))))
 
+(defn lint-as-aliased-usage
+  "Check if a namespace symbol has an existing `:as` alias."
+  [ctx ns-sym name-sym expr]
+  (let [expr (if (meta name-sym)
+               name-sym expr)]
+    (when-let [ns-sym (get (:aliases (:ns ctx)) ns-sym)]
+      (when (some-> ns-sym meta :alias meta :as-alias)
+        (findings/reg-finding!
+         ctx
+         (node->line
+          (:filename ctx)
+          expr
+          :aliased-namespace-var-usage
+          (format "Namespace only aliased but wasn't loaded: %s"
+                  ns-sym)))))))
+
 (defn resolve-name
   [ctx call? ns-name name-sym expr]
   (let [lang (:lang ctx)
@@ -527,6 +543,7 @@
                                (when (= (:name ns) ns-sym)
                                  ns-sym))]
               (lint-aliased-namespace ctx ns-sym name-sym expr)
+              (lint-as-aliased-usage ctx ns-sym name-sym expr)
               (let [core? (or (= 'clojure.core ns*)
                               (= 'cljs.core ns*))
                     [var-name interop] (if cljs?